backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
docker>=7.0.0,<8.0.0
loguru
biopython